requirements.txt
pymongo
python-dotenv
flask
bson
openai
beautifulsoup4
lxml
mysql-connector-python
